Create Problem2.java
Method should trigger another helper method that searches the array recursively for the value x. Your method should return the index of the element if it’s found (you may assume that each element is unique). Return -1 if the element is not found. Finally, create a main method that builds an array of String objects and then sort that array with Arrays.sort.