Create Fail2Ban.java
This program will take two command line arguments: the name of a single log file (such as the one attached to this assignment) and the name of an output file. You will read each line of the program in, extract the IP address from the line and determine whether the line is a failed login attempt or valid login attempt. Keep track of the number of times a particular IP address fails to login. Your program should then open the output file and print a list of IP addresses to that fail 3 or more times.
A sample command line execution of this file might look like this:

$ java Fail2Ban logs_processed.txt output.txt

The list of IP addresses that have failed 3 or more times should appear in the output.txt file.